[md] fix 'allow faster resync only on non-rotational media' underneath dm
Message-id: <1536851148-14583-3-git-send-email-ncroxon@redhat.com>
Patchwork-id: 228820
O-Subject: [RHEL7.6 V5 PATCH 2/2] md: fix 'allow faster resync only on non-rotational media' underneath dm
Bugzilla: 1561162
RH-Acked-by: Xiao Ni <xni@redhat.com>
RH-Acked-by: Jerry Snitselaar <jsnitsel@redhat.com>
RH-Acked-by: Heinz Mauelshagen <heinzm@redhat.com>

Tested: lvcreate of a raid5 LV and lvm2 test suite pass
Upstream status: RHEL-only

el7 commit 58cdca88daf2423248c8491f0a46d76dc3d9d0d7
introduced a regression to dm-raid on top of md-raid.

In this stacked case, MD doesn't manage the request queue thus
mddev->queue is NULL and must not be accessed unconditionally!

Why this is a RHEL-only patch. This issue is not seen upstream nor
in other distributions. We have a high value customer that
requries this change. We have presented this patch to upstream
and that discussion is on-going. To temporary help our customer
we want to release this patch that fixes their environment.
When an upstream resolution comes to fruit, we will work
that into RHEL.

Signed-off-by: Heinz Mauelshagen <heinzm@redhat.com>
Signed-off-by: Nigel Croxon <ncroxon@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>