[redhat] spec: fix BuildRequires for bpftool
Message-id: <20180919132627.27832-1-jmarchan@redhat.com>
Patchwork-id: 228993
O-Subject: [RHEL7.6 PATCH 2/1] spec: fix BuildRequires for bpftool
Bugzilla: 1598735
RH-Acked-by: Bruno Eduardo de Oliveira Meneguele <bmeneg@redhat.com>
RH-Acked-by: Eugene Syromiatnikov <esyr@redhat.com>
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: Luis Goncalves <lgoncalv@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1598735
Brew build: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=18395541

The previous patch put the BuildRequires in the wrong place. It was
actually part of the description which is not quite what it was meant
to be. This patch moves it in the proper place.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>