[fs] nfsdv4: use export cache flushtime for changeid on V4ROOT objects
Message-id: <20181008163124.20523-1-bfields@redhat.com>
Patchwork-id: 230074
O-Subject: [Patch RHEL7.7 fs] NFSDv4: use export cache flushtime for changeid on V4ROOT objects.
Bugzilla: 1637101 1570022
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>
RH-Acked-by: Scott Mayhew <smayhew@redhat.com>

From: NeilBrown <neilb@suse.com>

If you change the set of filesystems that are exported, then
the contents of various directories in the NFSv4 pseudo-root
is likely to change.  However the change-id of those
directories is currently tied to the underlying directory,
so the client may not see the changes in a timely fashion.

This patch changes the change-id number to be derived from the
"flush_time" of the export cache.  Whenever any changes are
made to the set of exported filesystems, this flush_time is
updated.  The result is that clients see changes to the set
of exported filesystems much more quickly, often immediately.

Signed-off-by: NeilBrown <neilb@suse.com>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
(cherry picked from commit b880092109323d1cc608c420fa712d3e19508e13)
Bugzilla: 1637101
Brew: 18686699
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>