[tools] perf script: Allow printing period for non freq mode groups
Message-id: <20181023074339.20119-120-jolsa@redhat.com>
Patchwork-id: 231464
O-Subject: [PATCH RHEL7.7 BZ1641373 119/294] perf script: Allow printing period for non freq mode groups
Bugzilla: 1641373
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1641373
https://bugzilla.redhat.com/show_bug.cgi?id=1641373

upstream
========
commit 5039c8a28fa97b8dce7b363a5ecd4bee2b87bf03
Author: Andi Kleen <ak@linux.intel.com>
Date: Thu Nov 9 06:55:26 2017 -0800

description
===========
When using leader sampling the values of the not sampled but counted
events are shown by perf script in "period".

Currently printing period is only allowed when the main event has a
period, that is it is in frequency mode.

This implies that we cannot dump the values of counted events when the
leader event is not in frequency mode.

Just remove the check that the period must be set on all events. It will
just be printed as 0 instead if it's not available.

This fixes the following:

  $ perf record -c 100000 -e '{cycles,branches}:S'
  $ perf script -F event,period

Further commentary by Jiri Olsa:

The period will be the value of configured period, not 0:

int perf_evsel__parse_sample(struct ...
  ...
  data->period = evsel->attr.sample_period;

  $ perf record -c 100000
  $ perf script -F event,period | head -3
  Failed to open /tmp/perf-2048.map, continuing without symbols
      100000 cycles:ppp:
      100000 cycles:ppp:

other than that I think we can remove that check, because we will have
always sane number in period

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>