[tools] perf rblist: Create rblist__exit() function
Message-id: <20181023074339.20119-149-jolsa@redhat.com>
Patchwork-id: 231538
O-Subject: [PATCH RHEL7.7 BZ1641373 148/294] perf rblist: Create rblist__exit() function
Bugzilla: 1641373
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1641373
https://bugzilla.redhat.com/show_bug.cgi?id=1641373

upstream
========
commit 33fec3e393dc1c55737cfb9c876b5c0da0d6f380
Author: Jin Yao <yao.jin@linux.intel.com>
Date: Fri Dec 1 18:57:25 2017 +0800

description
===========
Currently we have a rblist__delete() which is used to delete a rblist.
While rblist__delete() will free the pointer of rblist at the end.

It's an inconvenience for the user to delete a rblist which is not
allocated by something like malloc(). For example, the rblist is
embedded in a larger data structure.

This patch creates a new function rblist__exit() which is similar to
rblist__delete() but it will not free the pointer of rblist.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>