[tools] perf stat: Define a structure for per-thread shadow stats
Message-id: <20181023074339.20119-159-jolsa@redhat.com>
Patchwork-id: 231561
O-Subject: [PATCH RHEL7.7 BZ1641373 158/294] perf stat: Define a structure for per-thread shadow stats
Bugzilla: 1641373
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1641373
https://bugzilla.redhat.com/show_bug.cgi?id=1641373

upstream
========
commit e5fcc2abc353be94548080d84de3269ef6cc2af6
Author: Jin Yao <yao.jin@linux.intel.com>
Date: Tue Dec 5 22:03:01 2017 +0800

description
===========
Perf has a set of static variables to record the runtime shadow metrics
stats.

While if we want to record the runtime shadow stats for per-thread, it
will be the limitation. This patch creates a structure and the next
patches will use this structure to update the runtime shadow stats for
per-thread.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>