[tools] perf stat: Update per-thread shadow stats
Message-id: <20181023074339.20119-162-jolsa@redhat.com>
Patchwork-id: 231570
O-Subject: [PATCH RHEL7.7 BZ1641373 161/294] perf stat: Update per-thread shadow stats
Bugzilla: 1641373
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1641373
https://bugzilla.redhat.com/show_bug.cgi?id=1641373

upstream
========
commit 1fcd03946b52b8a57a6692fedd4406b45baedfe6
Author: Jin Yao <yao.jin@linux.intel.com>
Date: Tue Dec 5 22:03:04 2017 +0800

description
===========
The functions perf_stat__update_shadow_stats() is called to update the
shadow stats on a set of static variables.

But the static variables are the limitations to be extended to support
per-thread shadow stats.

This patch lets the perf_stat__update_shadow_stats() support to update
the shadow stats on a input parameter 'st' and uses
update_runtime_stat() to update the stats. It will not directly update
the static variables as before.

[ Rename 'stat' variables to 'st' to build on centos:{5,6} and others where it shadows a global declaration ]

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>