[tools] perf tools: Return all events as auto-completions after comma
Message-id: <20181023074339.20119-189-jolsa@redhat.com>
Patchwork-id: 231586
O-Subject: [PATCH RHEL7.7 BZ1641373 188/294] perf tools: Return all events as auto-completions after comma
Bugzilla: 1641373
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1641373
https://bugzilla.redhat.com/show_bug.cgi?id=1641373

upstream
========
commit 34c16db0f035f3f3dc50fbed03747693c12b6a5b
Author: Jin Yao <yao.jin@linux.intel.com>
Date: Fri Dec 22 18:57:35 2017 +0800

description
===========
It's a follow up for one previous patch "perf tool: Improve bash command
line auto-complete for multiple events with comma."

It fixes an issue that no events are displayed when <TAB> is directly
typed after comma.

With this patch, now the result is:

  root@skl:/tmp# perf stat -e cpu-cycles,<TAB>
  Display all 2389 possibilities? (y or n)
  alarmtimer:alarmtimer_cancel
  alarmtimer:alarmtimer_fired
  alarmtimer:alarmtimer_start
  alarmtimer:alarmtimer_suspend
  alignment-faults
  arith.divider_active
  BAClear_Cost
  baclears.any
  block:block_bio_backmerge
  block:block_bio_bounce
  block:block_bio_complete
  block:block_bio_frontmerge
  block:block_bio_queue
  block:block_bio_remap
  block:block_dirty_buffer
  block:block_getrq
  block:block_plug
  block:block_rq_complete
  block:block_rq_insert
  block:block_rq_issue
  block:block_rq_remap
  block:block_rq_requeue
  block:block_sleeprq
  --More--

One remaining issue is that the auto-completions doesn't work well
for the event with ':'. For example, clk:clk_enable.

Because ':' is set as WORDBREAK by default in bash. Need more work
for this case.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>