[tools] perf tools: Create function to perform multiple time range checking
Message-id: <20181023074339.20119-208-jolsa@redhat.com>
Patchwork-id: 231615
O-Subject: [PATCH RHEL7.7 BZ1641373 207/294] perf tools: Create function to perform multiple time range checking
Bugzilla: 1641373
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1641373
https://bugzilla.redhat.com/show_bug.cgi?id=1641373

upstream
========
commit 9a9b8b4b2271e763c1600311a3d4ecc2ac359b55
Author: Jin Yao <yao.jin@linux.intel.com>
Date: Fri Dec 8 21:13:44 2017 +0800

description
===========
Previous patch supports the multiple time range.

For example, select the first and second 10% time slices.
perf report --time 10%/1,10%/2

We need a function to check if a timestamp is in the ranges of
[0, 10%) and [10%, 20%].

Note that it includes the last element in [10%, 20%] but it doesn't
include the last element in [0, 10%). It's to avoid the overlap.

This patch implments a new function perf_time__ranges_skip_sample
for this checking.

Change log:

v4: Let perf_time__ranges_skip_sample be compatible with
    perf_time__skip_sample when only one time range.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>