[tools] tools lib traceevent: Simplify pointer print logic and fix %pF
Message-id: <20181023074339.20119-229-jolsa@redhat.com>
Patchwork-id: 231564
O-Subject: [PATCH RHEL7.7 BZ1641373 228/294] tools lib traceevent: Simplify pointer print logic and fix %pF
Bugzilla: 1641373
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1641373
https://bugzilla.redhat.com/show_bug.cgi?id=1641373

upstream
========
commit 38d70b7ca1769f26c0b79f3c08ff2cc949712b59
Author: Steven Rostedt (VMware) <rostedt@goodmis.org>
Date: Thu Jan 11 19:47:45 2018 -0500

description
===========
When processing %pX in pretty_print(), simplify the logic slightly by
incrementing the ptr to the format string if isalnum(ptr[1]) is true.
This follows the logic a bit more closely to what is in the kernel.

Also, this fixes a small bug where %pF was not giving the offset of the
function.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>