[tools] tools lib traceevent: Handle new pointer processing of bprint strings
Message-id: <20181023074339.20119-230-jolsa@redhat.com>
Patchwork-id: 231569
O-Subject: [PATCH RHEL7.7 BZ1641373 229/294] tools lib traceevent: Handle new pointer processing of bprint strings
Bugzilla: 1641373
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1641373
https://bugzilla.redhat.com/show_bug.cgi?id=1641373

upstream
========
commit 37db96bb49629681cb839d7304a70524fe10f969
Author: Steven Rostedt (VMware) <rostedt@goodmis.org>
Date: Thu Jan 11 19:47:46 2018 -0500

description
===========
The Linux kernel printf() has some extended use cases that dereference
the pointer. This is dangerouse for tracing because the pointer that is
dereferenced can change or even be unmapped. It also causes issues when
the trace data is extracted, because user space does not have access to
the contents of the pointer even if it still exists.

To handle this, the kernel was updated to process these dereferenced
pointers at the time they are recorded, and not post processed. Now they
exist in the tracing buffer, and no dereference is needed at the time of
reading the trace.

The event parsing library needs to handle this new case.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>