[tools] tools lib traceevent: Show contents (in hex) of data of unrecognized type records
Message-id: <20181023074339.20119-231-jolsa@redhat.com>
Patchwork-id: 231572
O-Subject: [PATCH RHEL7.7 BZ1641373 230/294] tools lib traceevent: Show contents (in hex) of data of unrecognized type records
Bugzilla: 1641373
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1641373
https://bugzilla.redhat.com/show_bug.cgi?id=1641373

upstream
========
commit e877372880f72399323e433187cce2bfbea40263
Author: Steven Rostedt (VMware) <rostedt@goodmis.org>
Date: Thu Jan 11 19:47:47 2018 -0500

description
===========
When a record has an unrecognized type, an error message is reported,
but it would also be helpful to see the contents of that record. At
least show what it is in hex, instead of just showing a blank line.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>