[tools] perf annotate browser: Be more robust when drawing jump arrows
Message-id: <20181023074339.20119-291-jolsa@redhat.com>
Patchwork-id: 231655
O-Subject: [PATCH RHEL7.7 BZ1641373 290/294] perf annotate browser: Be more robust when drawing jump arrows
Bugzilla: 1641373
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1641373
https://bugzilla.redhat.com/show_bug.cgi?id=1641373

upstream
========
commit 9cf195f80c5e8a6d779119a7d292e537315d2ea6
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Thu Mar 1 11:33:59 2018 -0300

description
===========
This first happened with a gcc function, _cpp_lex_token, that has the
usual jumps:

 1159e6c:  jne    115aa32 <_cpp_lex_token@@Base+0xf92>

I.e. jumps to a label inside that function (_cpp_lex_token), and those
works, but also this kind:

 1159e8b:  jne    c469be <cpp_named_operator2name@@Base+0xa72>

I.e. jumps to another function, outside _cpp_lex_token, which are not
being correctly handled generating as a side effect references to
ab->offset[] entries that are set to NULL, so to make this code more
robust, check that here.

A proper fix for will be put in place, looking at the function name
right after the '<' token and probably treating this like a 'call'
instruction.

For now just don't draw the arrow.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>