[tools] perf record: Fix crash in pipe mode
Message-id: <20181023074339.20119-292-jolsa@redhat.com>
Patchwork-id: 231649
O-Subject: [PATCH RHEL7.7 BZ1641373 291/294] perf record: Fix crash in pipe mode
Bugzilla: 1641373
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1641373
https://bugzilla.redhat.com/show_bug.cgi?id=1641373

upstream
========
commit cfacbabd1d9c35d2a179650b2911f17a8d8620b8
Author: Jiri Olsa <jolsa@kernel.org>
Date: Fri Mar 2 17:13:54 2018 +0100

description
===========
Currently we can crash perf record when running in pipe mode, like:

  $ perf record ls | perf report
  # To display the perf.data header info, please use --header/--header-only options.
  #
  perf: Segmentation fault
  Error:
  The - file has no samples!

The callstack of the crash is:

    0x0000000000515242 in perf_event__synthesize_event_update_name
  3513            ev = event_update_event__new(len + 1, PERF_EVENT_UPDATE__NAME, evsel->id[0]);
  (gdb) bt
  #0  0x0000000000515242 in perf_event__synthesize_event_update_name
  #1  0x00000000005158a4 in perf_event__synthesize_extra_attr
  #2  0x0000000000443347 in record__synthesize
  #3  0x00000000004438e3 in __cmd_record
  #4  0x000000000044514e in cmd_record
  #5  0x00000000004cbc95 in run_builtin
  #6  0x00000000004cbf02 in handle_internal_command
  #7  0x00000000004cc054 in run_argv
  #8  0x00000000004cc422 in main

The reason of the crash is that the evsel does not have ids array
allocated and the pipe's synthesize code tries to access it.

We don't force evsel ids allocation when we have single event, because
it's not needed. However we need it when we are in pipe mode even for
single event as a key for evsel update event.

Fixing this by forcing evsel ids allocation event for single event, when
we are in pipe mode.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>