[tools] perf stat: Fix CVS output format for non-supported counters
Message-id: <20181023074339.20119-293-jolsa@redhat.com>
Patchwork-id: 231664
O-Subject: [PATCH RHEL7.7 BZ1641373 292/294] perf stat: Fix CVS output format for non-supported counters
Bugzilla: 1641373
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1641373
https://bugzilla.redhat.com/show_bug.cgi?id=1641373

upstream
========
commit 40c21898ba5372c14ef71717040529794a91ccc2
Author: Ilya Pronin <ipronin@twitter.com>
Date: Mon Mar 5 22:43:53 2018 -0800

description
===========
When printing stats in CSV mode, 'perf stat' appends extra separators
when a counter is not supported:

<not supported>,,L1-dcache-store-misses,mesos/bd442f34-2b4a-47df-b966-9b281f9f56fc,0,100.00,,,,

Which causes a failure when parsing fields. The numbers of separators
should be the same for each line, no matter if the counter is or not
supported.

Fixes: 92a61f6412d3 ("perf stat: Implement CSV metrics output")
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>