[kernel] bpf: add faked "ending" subprog
Message-id: <20181023114647.12260-25-jolsa@redhat.com>
Patchwork-id: 231673
O-Subject: [PATCH RHEL7.7 BZ1640092 24/68] bpf: add faked "ending" subprog
Bugzilla: 1640092
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>

Bugzilla: 1640092
https://bugzilla.redhat.com/show_bug.cgi?id=1640092

upstream
========
commit 4cb3d99c84ccbf728ff0e381c7c9815c3fa2bd5e
Author: Jiong Wang <jiong.wang@netronome.com>
Date: Wed May 2 16:17:19 2018 -0400

description
===========
There are quite a few code snippet like the following in verifier:

       subprog_start = 0;
       if (env->subprog_cnt == cur_subprog + 1)
               subprog_end = insn_cnt;
       else
               subprog_end = env->subprog_info[cur_subprog + 1].start;

The reason is there is no marker in subprog_info array to tell the end of
it.

We could resolve this issue by introducing a faked "ending" subprog.
The special "ending" subprog is with "insn_cnt" as start offset, so it is
serving as the end mark whenever we iterate over all subprogs.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>