[kernel] bpf: support 64-bit offsets for bpf function calls
Message-id: <20181023114647.12260-30-jolsa@redhat.com>
Patchwork-id: 231691
O-Subject: [PATCH RHEL7.7 BZ1640092 29/68] bpf: support 64-bit offsets for bpf function calls
Bugzilla: 1640092
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>

Bugzilla: 1640092
https://bugzilla.redhat.com/show_bug.cgi?id=1640092

upstream
========
commit 2162fed49fa86cb3b5c296565837e674c53cc7ae
Author: Sandipan Das <sandipan@linux.vnet.ibm.com>
Date: Thu May 24 12:26:45 2018 +0530

description
===========
The imm field of a bpf instruction is a signed 32-bit integer.
For JITed bpf-to-bpf function calls, it holds the offset of the
start address of the callee's JITed image from __bpf_call_base.

For some architectures, such as powerpc64, this offset may be
as large as 64 bits and cannot be accomodated in the imm field
without truncation.

We resolve this by:

[1] Additionally using the auxiliary data of each function to
    keep a list of start addresses of the JITed images for all
    functions determined by the verifier.

[2] Retaining the subprog id inside the off field of the call
    instructions and using it to index into the list mentioned
    above and lookup the callee's address.

To make sure that the existing JIT compilers continue to work
without requiring changes, we keep the imm field as it is.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>