[kernel] bpf/verifier: introduce BPF_PTR_TO_MAP_VALUE
Message-id: <20181023114647.12260-48-jolsa@redhat.com>
Patchwork-id: 231732
O-Subject: [PATCH RHEL7.7 BZ1640092 47/68] bpf/verifier: introduce BPF_PTR_TO_MAP_VALUE
Bugzilla: 1640092
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>

Bugzilla: 1640092
https://bugzilla.redhat.com/show_bug.cgi?id=1640092

upstream
========
commit 3e6a4b3e0289dc9540a2c1d8a20657f4707fbabb
Author: Roman Gushchin <guro@fb.com>
Date: Thu Aug 2 14:27:22 2018 -0700

description
===========
BPF_MAP_TYPE_CGROUP_STORAGE maps are special in a way
that the access from the bpf program side is lookup-free.
That means the result is guaranteed to be a valid
pointer to the cgroup storage; no NULL-check is required.

This patch introduces BPF_PTR_TO_MAP_VALUE return type,
which is required to cause the verifier accept programs,
which are not checking the map value pointer for being NULL.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>