[tools] bpftool: make it easier to feed hex bytes to bpftool
Message-id: <20181023114647.12260-66-jolsa@redhat.com>
Patchwork-id: 231709
O-Subject: [PATCH RHEL7.7 BZ1640092 65/68] tools: bpftool: make it easier to feed hex bytes to bpftool
Bugzilla: 1640092
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>

Bugzilla: 1640092
https://bugzilla.redhat.com/show_bug.cgi?id=1640092

upstream
========
commit 0c90f2243ec67eeacf9624ae52ab43c734fe0e93
Author: Quentin Monnet <quentin.monnet@netronome.com>
Date: Tue Apr 17 19:46:34 2018 -0700

description
===========
bpftool uses hexadecimal values when it dumps map contents:

    # bpftool map dump id 1337
    key: ff 13 37 ff  value: a1 b2 c3 d4 ff ff ff ff
    Found 1 element

In order to lookup or update values with bpftool, the natural reflex is
then to copy and paste the values to the command line, and to try to run
something like:

    # bpftool map update id 1337 key ff 13 37 ff \
            value 00 00 00 00 00 00 1a 2b
    Error: error parsing byte: ff

bpftool complains, because it uses strtoul() with a 0 base to parse the
bytes, and that without a "0x" prefix, the bytes are considered as
decimal values (or even octal if they start with "0").

To feed hexadecimal values instead, one needs to add "0x" prefixes
everywhere necessary:

    # bpftool map update id 1337 key 0xff 0x13 0x37 0xff \
            value 0 0 0 0 0 0 0x1a 0x2b

To make it easier to use hexadecimal values, add an optional "hex"
keyword to put after "key" or "value" to tell bpftool to consider the
digits as hexadecimal. We can now do:

    # bpftool map update id 1337 key hex ff 13 37 ff \
            value hex 0 0 0 0 0 0 1a 2b

Without the "hex" keyword, the bytes are still parsed according to
normal integer notation (decimal if no prefix, or hexadecimal or octal
if "0x" or "0" prefix is used, respectively).

The patch also add related documentation and bash completion for the
"hex" keyword.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>