[kernel] sched/cputime: Move the vtime task fields to their own struct
Message-id: <20181113103014.22883-4-yauheni.kaliuta@redhat.com>
Patchwork-id: 232828
O-Subject: [RHEL7.7 PATCH 3/4] sched/cputime: Move the vtime task fields to their own struct
Bugzilla: 1436351
RH-Acked-by: Luiz Capitulino <lcapitulino@redhat.com>
RH-Acked-by: Phil Auld <pauld@redhat.com>
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>

Bugzilla: https://bugzilla.redhat.com/1436351

This is a backport of

commit bac5b6b6b11560f323e71d0ebac4061cfe5f56c0
Author: Frederic Weisbecker <fweisbec@gmail.com>
Date:   Thu Jun 29 19:15:10 2017 +0200

    sched/cputime: Move the vtime task fields to their own struct

    We are about to add vtime accumulation fields to the task struct. Let's
    avoid more bloatification and gather vtime information to their own
    struct.

    Tested-by: Luiz Capitulino <lcapitulino@redhat.com>
    Signed-off-by: Frederic Weisbecker <fweisbec@gmail.com>
    Reviewed-by: Thomas Gleixner <tglx@linutronix.de>
    Acked-by: Rik van Riel <riel@redhat.com>
    Cc: Linus Torvalds <torvalds@linux-foundation.org>
    Cc: Peter Zijlstra <peterz@infradead.org>
    Cc: Wanpeng Li <kernellwp@gmail.com>
    Link: http://lkml.kernel.org/r/1498756511-11714-5-git-send-email-fweisbec@gmail.com
    Signed-off-by: Ingo Molnar <mingo@kernel.org>

Conflicts: missing b7ce2277f087 sched/cputime: Convert vtime_seqlock to seqcount
missing https://lkml.org/lkml/2017/1/29/223

Reimplementing the original patch logic: move the fields to the own
structure and change the corresponded access code.

KABI: struct task_struct allocated in the core, so it ok to put the
vtime structure to the end. For this patch it has the same size as
the replaced fields, but will be changed in the next patch.

Signed-off-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>