[mm] userfaultfd: shmem: allocate anonymous memory for MAP_PRIVATE shmem
Message-id: <20181127044744.21081-10-aarcange@redhat.com>
Patchwork-id: 233745
O-Subject: [RHEL7.7 PATCH 09/13] userfaultfd: shmem: allocate anonymous memory for MAP_PRIVATE shmem
Bugzilla: 1640519
CVE: CVE-2018-18397
RH-Acked-by: Peter Xu <peterx@redhat.com>
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>

Userfaultfd did not create private memory when UFFDIO_COPY was invoked
on a MAP_PRIVATE shmem mapping. Instead it wrote to the shmem file,
even when that had not been opened for writing. Though, fortunately,
that could only happen where there was a hole in the file.

Fix the shmem-backed implementation of UFFDIO_COPY to create private
memory for MAP_PRIVATE mappings. The hugetlbfs-backed implementation
was already correct.

This change is visible to userland, if userfaultfd has been used in
unintended ways: so it introduces a small risk of incompatibility, but
is necessary in order to respect file permissions.

An app that uses UFFDIO_COPY for anything like postcopy live migration
won't notice the difference, and in fact it'll run faster because
there will be no copy-on-write and memory waste in the tmpfs pagecache
anymore.

Userfaults on MAP_PRIVATE shmem keep triggering only on file holes
like before.

The real zeropage can also be built on a MAP_PRIVATE shmem mapping
through UFFDIO_ZEROPAGE and that's safe because the zeropage pte is
never dirty, in turn even an mprotect upgrading the vma permission
from PROT_READ to PROT_READ|PROT_WRITE won't make the zeropage pte
writable.

RHEL-only: in -mm for BZ 1640519
Reported-by: Mike Rapoport <rppt@linux.ibm.com>
Reviewed-by: Hugh Dickins <hughd@google.com>
Cc: stable@vger.kernel.org
Fixes: 4c27fe4c4c84 ("userfaultfd: shmem: add shmem_mcopy_atomic_pte for userfaultfd support")
Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>