[mm] userfaultfd: shmem: uffdio_copy: set the page dirty if VM_WRITE is not set
Message-id: <20181127044744.21081-13-aarcange@redhat.com>
Patchwork-id: 233751
O-Subject: [RHEL7.7 PATCH 12/13] userfaultfd: shmem: UFFDIO_COPY: set the page dirty if VM_WRITE is not set
Bugzilla: 1640519
CVE: CVE-2018-18397
RH-Acked-by: Peter Xu <peterx@redhat.com>
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>

Set the page dirty if VM_WRITE is not set because in such case the pte
won't be marked dirty and the page would be reclaimed without
writepage (i.e. swapout in the shmem case).

This was found by source review. Most apps (certainly including QEMU)
only use UFFDIO_COPY on PROT_READ|PROT_WRITE mappings or the app can't
modify the memory in the first place. This is for correctness and it
could help the non cooperative use case to avoid unexpected data loss.

RHEL-only: in -mm along with the other patches for BZ 1640519
Reviewed-by: Hugh Dickins <hughd@google.com>
Cc: stable@vger.kernel.org
Fixes: 4c27fe4c4c84 ("userfaultfd: shmem: add shmem_mcopy_atomic_pte for userfaultfd support")
Reported-by: Hugh Dickins <hughd@google.com>
Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>