[scsi] lpfc: Adding ability to reset chip via pci bus reset
Message-id: <1545401521-4327-61-git-send-email-dkennedy@redhat.com>
Patchwork-id: 236326
O-Subject: [RHEL7.7 PATCH e-stor 60/61] scsi: lpfc: Adding ability to reset chip via pci bus reset
Bugzilla: 1642905
RH-Acked-by: Ewan Milne <emilne@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: James Smart <jsmart2021@gmail.com>

This patch adds a "pci_bus_reset" option to the board_mode sysfs attribute.
This option uses the pci_reset_bus() api to reset the PCIe link the adapter
is on, which will reset the chip/adapter.  Prior to issuing this option,
all functions on the same chip must be placed in the offline state by the
admin. After the reset, all of the instances may be brought online again.

The primary purpose of this functionality is to support cases where
firmware update required a chip reset but the admin did not want to reboot
the machine in order to instantiate the firmware update.

Sanity checks take place prior to the reset to ensure the adapter is the
sole entity on the PCIe bus and that all functions are in the offline
state.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>

Conflicts:
a/drivers/scsi/lpfc/lpfc_attr.c
Had to modify lpfc_attr because the upstream api for pci_reset_bus
takes the pdev pointer as its parameter.

Signed-off-by: Dick Kennedy <dick.kennedy@broadcom.com>
Signed-off-by: James Smart <jsmart2021@gmail.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
(cherry picked from commit 5021267af1327724a6f1859bb1594660a4c22bd3)
Signed-off-by: Dick Kennedy <dkennedy@redhat.com>