[netdrv] nfp: flower: support offloading multiple rules with same cookie
Message-id: <1548782282-13140-4-git-send-email-pablo.cascon@redhat.com>
Patchwork-id: 239125
O-Subject: [RHEL7.7 BZ1637991 003/133] nfp: flower: support offloading multiple rules with same cookie
Bugzilla: 1637991
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1637991
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=20009692

When multiple netdevs are attached to a tc offload block and register for
callbacks, a rule added to the block will be propogated to all netdevs.
Previously these were detected as duplicates (based on cookie) and
rejected. Modify the rule nfp lookup function to optionally include an
ingress netdev and a host context along with the cookie value when
searching for a rule. When a new rule is passed to the driver, the netdev
the rule is to be attached to is considered when searching for dublicates.
When a stats update is received from HW, the host context is used
alongside the cookie to map to the correct host rule.

Signed-off-by: John Hurley <john.hurley@netronome.com>
Reviewed-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 54a4a03439305f701834cef1e4bd108548388d83)
Signed-off-by: Pablo Cascn <pablo.cascon@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>