[x86] perf/x86/intel: Properly save/restore the PMU state in the NMI handler
Message-id: <20190211232742.16489-3-jolsa@redhat.com>
Patchwork-id: 241401
O-Subject: [PATCH RHEL7 BZ1672651 002/271] perf/x86/intel: Properly save/restore the PMU state in the NMI handler
Bugzilla: 1672651
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1672651
https://bugzilla.redhat.com/show_bug.cgi?id=1672651

upstream
========
commit 82d71ed0277efc45360828af8c4e4d40e1b45352
Author: Kan Liang <kan.liang@linux.intel.com>
Date: Tue Feb 20 02:11:50 2018 -0800

description
===========
The PMU is disabled in intel_pmu_handle_irq(), but cpuc->enabled is not updated
accordingly.

This is fine in current usage because no-one checks it - but fix it
for future code: for example, the drain_pebs() will be modified to
fix an auto-reload bug.

Properly save/restore the old PMU state.

Signed-off-by: Jan Stancek <jstancek@redhat.com>