[tools] perf kmem: Document a missing option & an argument
Message-id: <20190211232742.16489-34-jolsa@redhat.com>
Patchwork-id: 241437
O-Subject: [PATCH RHEL7 BZ1672651 033/271] perf kmem: Document a missing option & an argument
Bugzilla: 1672651
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1672651
https://bugzilla.redhat.com/show_bug.cgi?id=1672651

upstream
========
commit 577980a00063935815a55f461601579fd5e61f59
Author: Sangwon Hong <qpakzk@gmail.com>
Date: Mon Feb 12 05:38:36 2018 +0900

description
===========
First, 'perf kmem' has a '--force' option, but didn't document it on the
man page. So add it.

Second, the '--time' option has to get a value, but isn't documented on
the man page. Describe it.

[ Add blank like after --force block, as requested by Namhyung ]

Signed-off-by: Jan Stancek <jstancek@redhat.com>