[tools] perf machine: Fix paranoid check in machine__set_kernel_mmap()
Message-id: <20190211232742.16489-47-jolsa@redhat.com>
Patchwork-id: 241432
O-Subject: [PATCH RHEL7 BZ1672651 046/271] perf machine: Fix paranoid check in machine__set_kernel_mmap()
Bugzilla: 1672651
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1672651
https://bugzilla.redhat.com/show_bug.cgi?id=1672651

upstream
========
commit 1d12cec6ce99614297e10945d917fd8a62cd2b09
Author: Namhyung Kim <namhyung@kernel.org>
Date: Mon Feb 19 19:00:46 2018 +0900

description
===========
The machine__set_kernel_mmap() is to setup addresses of the kernel map
using external info.  But it has a check when the address is given from
an incorrect input which should have the start and end address of 0
(i.e. machine__process_kernel_mmap_event).

But we also use the end address of 0 for a valid input so change it to
check both start and end addresses.

Signed-off-by: Jan Stancek <jstancek@redhat.com>