[tools] perf cgroup: Rename close_cgroup() to cgroup__put()
Message-id: <20190211232742.16489-62-jolsa@redhat.com>
Patchwork-id: 241465
O-Subject: [PATCH RHEL7 BZ1672651 061/271] perf cgroup: Rename close_cgroup() to cgroup__put()
Bugzilla: 1672651
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1672651
https://bugzilla.redhat.com/show_bug.cgi?id=1672651

upstream
========
commit a53b646030ee6f65accdc49e772823b8134a37f6
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Tue Mar 6 10:10:45 2018 -0300

description
===========
It is not really closing the cgroup, but instead dropping a reference
count and if it hits zero, then calling delete, which will, among other
cleanup shores, close the cgroup fd.

So it is really dropping a reference to that cgroup, and the method name
for that is "put", so rename close_cgroup() to cgroup__put() to follow
this naming convention.

Signed-off-by: Jan Stancek <jstancek@redhat.com>